Factorize container
